refactor(schema)!: replace `_schema`/`types` with a `schema` package (facade + private `_models`)
