+++ docs/api/schema.md
@@ -2,6 +2,6 @@
 
 The JSON Schema document model — what you build, or validate raw input into, before conversion.
 
-::: pydantic_jsonschema.types
+::: pydantic_jsonschema.schema
     options:
       show_if_no_docstring: true
