+++ pydantic_jsonschema/converters/_metadata.py
@@ -1,6 +1,6 @@
 """`Annotated` metadata helpers for array / object value annotations."""
 
-# NOTE: `Schema` fields use `X | MISSING` unions (see `_schema.py`). mypy doesn't
+# NOTE: `Schema` fields use `X | MISSING` unions (see `schema/_models.py`). mypy doesn't
 # recognize `MISSING` as a type, so it infers fields without the `Sentinel` branch
 # and flags every `is not MISSING` check as a non-overlapping identity comparison.
 # mypy: disable-error-code="comparison-overlap"
@@ -11,7 +11,7 @@ import annotated_types
 from pydantic import AfterValidator
 from pydantic.experimental.missing_sentinel import MISSING
 
-from pydantic_jsonschema.types import Schema
+from pydantic_jsonschema.schema import Schema
 
 __all__ = [
     "annotate",
