+++ pydantic_jsonschema/converters/_object_keywords.py
@@ -6,7 +6,7 @@ converter-aware ones (`dependentSchemas` / `patternProperties` / `propertyNames`
 `SchemaConverter`.
 """
 
-# NOTE: `Schema` fields use `X | MISSING` unions (see `_schema.py`). mypy doesn't
+# NOTE: `Schema` fields use `X | MISSING` unions (see `schema/_models.py`). mypy doesn't
 # recognize `MISSING` as a type, so it infers fields without the `Sentinel` branch
 # and flags every `is not MISSING` check as a non-overlapping identity comparison.
 # mypy: disable-error-code="comparison-overlap"
@@ -16,7 +16,7 @@ from typing import Any
 from pydantic import model_validator
 from pydantic.experimental.missing_sentinel import MISSING
 
-from pydantic_jsonschema.types import Schema
+from pydantic_jsonschema.schema import Schema
 
 from ._utils import unwrap
