+++ pydantic_jsonschema/schema/__init__.py
@@ -1,0 +1,17 @@
+"""Public JSON Schema models: `Schema`, `Reference`, and the `DataType` enum.
+
+The models live in the private `_models` submodule; this package re-exports only the public
+names. See [the schema guide](https://danipulok.github.io/pydantic-jsonschema/latest/schema/).
+"""
+
+from pydantic_jsonschema.schema._models import (
+    DataType,
+    Reference,
+    Schema,
+)
+
+__all__ = [
+    "DataType",
+    "Reference",
+    "Schema",
+]
