+++ pydantic_jsonschema/schema/_models.py
