--- pydantic_jsonschema/types.py
@@ -1,13 +1,0 @@
-"""Public JSON Schema models: `Schema`, `Reference`, and the `DataType` enum."""
-
-from pydantic_jsonschema._schema import (
-    DataType,
-    Reference,
-    Schema,
-)
-
-__all__ = [
-    "DataType",
-    "Reference",
-    "Schema",
-]
