+++ pyproject.toml
@@ -159,7 +159,7 @@ ignore = [
 "pydantic_jsonschema/formats/_types.py" = [
     "E501",     # Long spec URLs in format-alias attribute docstrings
 ]
-"pydantic_jsonschema/_schema.py" = [
+"pydantic_jsonschema/schema/_models.py" = [
     "E501",     # Long spec URLs in `Schema` field attribute docstrings
 ]
 "tests/formats/test_hostname.py" = [
