+++ tests/applicators/test_one_of.py
@@ -5,7 +5,7 @@ from inline_snapshot import snapshot
 from pydantic import ValidationError
 
 from pydantic_jsonschema import to_model
-from pydantic_jsonschema.types import Schema
+from pydantic_jsonschema.schema import Schema
 from tests.conftest import dump_errors
 
 __all__: list[str] = []
